+++ database/sql/src/main/scala/cromwell/database/slick/tables/CallCachingSimpletonEntryComponent.scala
@@ -20,7 +20,7 @@ trait CallCachingSimpletonEntryComponent {
 
     def wdlType = column[String]("WDL_TYPE", O.Length(255))
 
-    def callCachingEntryId = column[Int]("CALL_CACHING_ENTRY_ID")
+    def callCachingEntryId = column[Long]("CALL_CACHING_ENTRY_ID")
 
     override def * = (simpletonKey, simpletonValue, wdlType, callCachingEntryId.?, callCachingSimpletonEntryId.?) <>
       (CallCachingSimpletonEntry.tupled, CallCachingSimpletonEntry.unapply)
@@ -41,7 +41,7 @@ trait CallCachingSimpletonEntryComponent {
     * Find all result simpletons which match a given CALL_CACHING_ENTRY_ID
     */
   val callCachingSimpletonEntriesForCallCachingEntryId = Compiled(
-    (callCachingEntryId: Rep[Int]) => for {
+    (callCachingEntryId: Rep[Long]) => for {
       callCachingSimpletonEntry <- callCachingSimpletonEntries
       if callCachingSimpletonEntry.callCachingEntryId === callCachingEntryId
     } yield callCachingSimpletonEntry
