+++ database/sql/src/main/scala/cromwell/database/sql/tables/CallCachingSimpletonEntry.scala
@@ -7,6 +7,6 @@ case class CallCachingSimpletonEntry
   simpletonKey: String,
   simpletonValue: Option[SerialClob],
   wdlType: String,
-  callCachingEntryId: Option[Int] = None,
+  callCachingEntryId: Option[Long] = None,
   callCachingSimpletonEntryId: Option[Int] = None
 )
